Renaming App.jsx for troubleshooting
