+++ frontEnd/main/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App.jsx'
+import App from './Main.jsx'
 
 ReactDOM.render(
 <App />,
